+++ src/parse/expr/parse_expr.rs
@@ -4,11 +4,11 @@ use crate::parse::resp;
 use crate::parse::expr::parse_trim::parse_trim;
 
 
-pub fn parse_expr(tokens: Vec<token::Token>, index: &mut i32) -> node::CalculNode {
+pub fn parse_expr(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculNode {
     let mut node = parse_trim(tokens.clone(), index);
 
-    while tokens.len() > *index as usize {
-        match tokens[*index as usize].kind {
+    while tokens.len() > *index {
+        match tokens[*index].kind {
             token::TokenKind::TokenAdd => {
                 *index += 1;
                 let right = parse_trim(tokens.clone(), index);
