+++ src/parse/expr/parse_trim.rs
@@ -4,11 +4,11 @@ use crate::parse::expr::parse_factor::parse_factor;
 use crate::parse::resp;
 
 
-pub fn parse_trim(tokens: Vec<token::Token>, index: &mut i32) -> node::CalculNode {
+pub fn parse_trim(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculNode {
     let mut node = parse_factor(tokens.clone(), index);
 
-    while tokens.len() > *index as usize {
-        match tokens[*index as usize].kind {
+    while tokens.len() > *index {
+        match tokens[*index].kind {
             token::TokenKind::TokenMul => {
                 *index += 1;
                 let right = parse_factor(tokens.clone(), index);
