+++ src/parse/semantic/cond_branch.rs
@@ -4,11 +4,11 @@ use crate::parse::resp;
 use crate::parse::expr::parse_comper::parse_comper_op;
 
 
-pub fn make_if_node(tokens: Vec<token::Token>, index: &mut i32) -> node::CalculNode {
+pub fn make_if_node(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculNode {
     let mut if_node = resp::handler::make_null_node();
 
-    while tokens.len() > *index as usize {
-        match tokens[*index as usize].kind {
+    while tokens.len() > *index {
+        match tokens[*index].kind {
             token::TokenKind::TokenSpace => {},
             token::TokenKind::TokenIf => if_node.node_type = node::NodeKind::NodeIf,
             token::TokenKind::TokenLBrace => {
@@ -25,12 +25,12 @@ pub fn make_if_node(tokens: Vec<token::Token>, index: &mut i32) -> node::CalculN
     if_node
 }
 
-pub fn make_if_else_node(tokens: Vec<token::Token>, index: &mut i32) -> Option<node::CalculNode> {
+pub fn make_if_else_node(tokens: Vec<token::Token>, index: &mut usize) -> Option<node::CalculNode> {
     let mut if_else_node = resp::handler::make_null_node();
     if_else_node.node_type = node::NodeKind::NodeIfElse;
 
-    while tokens.len() > *index as usize {
-        match tokens[*index as usize].kind {
+    while tokens.len() > *index {
+        match tokens[*index].kind {
             token::TokenKind::TokenSpace => {},
             token::TokenKind::TokenElse => {
                 return Some(make_else_node(tokens.clone(), index))
@@ -53,11 +53,11 @@ pub fn make_if_else_node(tokens: Vec<token::Token>, index: &mut i32) -> Option<n
     Some(if_else_node)
 }
 
-pub fn make_else_node(tokens: Vec<token::Token>, index: &mut i32) -> node::CalculNode {
+pub fn make_else_node(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculNode {
     let mut else_node = resp::handler::make_null_node();
 
-    while tokens.len() > *index as usize {
-        match tokens[*index as usize].kind {
+    while tokens.len() > *index {
+        match tokens[*index].kind {
             token::TokenKind::TokenSpace => {},
             token::TokenKind::TokenElse => else_node = resp::handler::make_true_node(node::NodeKind::NodeElse),
             token::TokenKind::TokenLBrace => {
